+++ src/main/scala/dxWDL/Main.scala
@@ -1,6 +1,7 @@
 package dxWDL
 
-import com.dnanexus.{DXProject}
+import com.dnanexus._
+import com.fasterxml.jackson.databind.JsonNode
 import com.typesafe.config._
 import java.nio.file.{Path, Paths}
 import scala.collection.mutable.HashMap
@@ -627,6 +628,38 @@ object Main extends App {
         SuccessfulTermination(s"success ${op}")
     }
 
+
+    // Get the WOM source code, and the instance type database from the
+    // details field stored on the platform
+    private def retrieveFromDetails(jobInfoPath: Path) : (String, InstanceTypeDB) = {
+        val jobInfo = Utils.readFileContent(jobInfoPath).parseJson
+        val applet: DXApplet = jobInfo.asJsObject.fields.get("applet") match {
+            case None =>
+                Utils.trace(true,
+                            s"""|applet field not found locally, performing
+                                |an API call.
+                                |""".stripMargin)
+                val dxJob : DXJob = Utils.dxEnv.getJob()
+                dxJob.describe().getApplet()
+            case Some(JsString(x)) =>
+                DXApplet.getInstance(x)
+            case Some(other) =>
+                throw new Exception(s"malformed applet field ${other} in job info")
+        }
+
+        val descOptions = DXDataObject.DescribeOptions.get().withDetails
+        val details: JsValue = Utils.jsValueOfJsonNode(
+            applet.describe(descOptions).getDetails(classOf[JsonNode]))
+
+        val JsString(womSourceCodeEncoded) = details.asJsObject.fields("womSourceCode")
+        val womSourceCode = Utils.base64DecodeAndGunzip(womSourceCodeEncoded)
+
+        val JsString(instanceTypeDBEncoded) = details.asJsObject.fields("instanceTypeDB")
+        val dbRaw = Utils.base64DecodeAndGunzip(instanceTypeDBEncoded)
+        val instanceTypeDB = dbRaw.parseJson.convertTo[InstanceTypeDB]
+        (womSourceCode, instanceTypeDB)
+    }
+
     def internalOp(args : Seq[String]) : Termination = {
         val operation = InternalOp.values find (x => normKey(x.toString) == normKey(args.head))
         operation match {
@@ -635,20 +668,14 @@ object Main extends App {
             case Some(op) if args.length == 3 =>
                 val homeDir = Paths.get(args(1))
                 val rtDebugLvl = parseRuntimeDebugLevel(args(2))
-                val (jobInputPath, jobOutputPath, jobErrorPath, _) =
+                val (jobInputPath, jobOutputPath, jobErrorPath, jobInfoPath) =
                     Utils.jobFilesOfHomeDir(homeDir)
                 val dxPathConfig = buildRuntimePathConfig(rtDebugLvl >= 1)
                 val dxIoFunctions = DxIoFunctions(dxPathConfig, rtDebugLvl)
 
-                // Figure out the available instance types, and their prices,
-                // by reading the file
-                val dbRaw = Utils.readFileContent(dxPathConfig.instanceTypeDB)
-                val instanceTypeDB = dbRaw.parseJson.convertTo[InstanceTypeDB]
-
-                   // Get the WOM source code (currently WDL, could be also CWL in the future)
-                var womSourceCodeEncoded = Utils.readFileContent(dxPathConfig.womSourceCodeEncoded)
-                womSourceCodeEncoded = womSourceCodeEncoded.substring(0, womSourceCodeEncoded.length - 1)
-                val womSourceCode = Utils.base64DecodeAndGunzip(womSourceCodeEncoded)
+                // Get the WOM source code (currently WDL, could be also CWL in the future)
+                // Parse the inputs, convert to WOM values.
+                val (womSourceCode, instanceTypeDB) = retrieveFromDetails(jobInfoPath)
 
                 try {
                     op match {