+++ RELEASE_NOTES.md
@@ -3,10 +3,11 @@
 ## 1.02
 - Removed the instance-type database, and wom source code from the inputs of applets.
 - Added the WDL source code to workflow and applet objects on the platform. It is stored in the details field, and
-can be easily [retrieved](./doc/ExpertOptions.md#Getting-WDL-sources).
-- Bug fixes
+can be easily [retrieved](./doc/ExpertOptions.md#Getting-WDL-sources). It has been removed from the generated applet bash script.
 - Initial support for the `struct` type
-- Bug fixes: https://github.com/dnanexus/dxWDL/issues/224, https://github.com/dnanexus/dxWDL/issues/227
+- Check that the reserved substring '___' is not used in the source WDL code. This sequence is used
+  to translate dots ('.') into DNAx inputs and outputs. Dots are invalid symbols there.
+- Bug fixes: https://github.com/dnanexus/dxWDL/issues/224, https://github.com/dnanexus/dxWDL/issues/227, https://github.com/dnanexus/dxWDL/issues/228
 
 ## 1.01
 - Ensure that native docker uses the machine's hostname (i.e., the job ID) as